requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
appdirs==1.4.4
astroid==2.15.5
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
clickhouse-connect==0.5.25
coloredlogs==15.0.1
comm==0.1.3
confection==0.0.4
cryptography==41.0.1
cymem==2.0.7
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.6
dnspython==2.3.0
docopt==0.6.2
duckdb==0.8.0
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.95.2
filelock==3.12.0
Flask==2.3.2
Flask-Cors==3.0.10
flatbuffers==23.5.9
frozenlist==1.3.3
fsspec==2023.5.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
httptools==0.5.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.23.1
ipython==8.13.2
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
langchain==0.0.178
langcodes==3.3.0
lazy-object-proxy==1.9.0
lxml==4.9.2
lz4==4.3.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
onnxruntime==1.15.0
openai==0.27.7
openapi-schema-pydantic==1.2.4
outcome==1.2.0
overrides==7.3.1
packaging==23.1
pandas==2.0.1
parso==0.8.3
pathy==0.10.1
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.0
platformdirs==3.5.1
posthog==3.0.1
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.8
pydeck==0.8.1b0
pyee==8.2.2
pyflakes==3.0.1
PyGithub==1.58.2
Pygments==2.15.1
PyJWT==2.7.0
pylint==2.17.4
pymongo==4.3.3
Pympler==1.0.1
PyNaCl==1.5.0
pypdf==3.9.0
PyPDF2==3.0.1
pyppeteer==1.0.2
pyreadline3==3.4.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
regex==2023.5.5
requests==2.31.0
rich==13.3.5
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.9.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.48
srsly==2.4.6
stack-data==0.6.2
starlette==0.27.0
streamlit==1.22.0
sympy==1.12
tenacity==8.2.2
thinc==8.1.10
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
tomlkit==0.11.8
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.2
trio==0.22.0
trio-websocket==0.10.2
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.6.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
uvicorn==0.22.0
validators==0.20.0
wasabi==1.1.1
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==10.4
Werkzeug==2.3.4
wrapt==1.15.0
wsproto==1.2.0
yarg==0.1.9
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
gunicorn==20.1.0
waitress==2.1.2